+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v5.3.1
 
+* `fixed` Allow `client.auto()` being called with an empty CSR common name
 * `fixed` Bug when calling `updateAccountKey()` with external account binding
 
 ## v5.3.0 (2024-02-05)